refactor: app 모듈에 있는 모든 컨트롤러를 어노테이션 기반 MVC로 변경
