+++ app/src/main/java/com/techcourse/domain/User.java
@@ -22,6 +22,10 @@ public class User {
         return account;
     }
 
+    public String getEmail() {
+        return email;
+    }
+
     @Override
     public String toString() {
         return "User{" +
