+++ app/src/main/webapp/login.jsp
@@ -30,7 +30,7 @@
                                         </form>
                                     </div>
                                     <div class="card-footer text-center py-3">
-                                        <div class="small"><a href="/register/view">아이디가 없나요? 회원가입 하러가기</a></div>
+                                        <div class="small"><a href="/register">아이디가 없나요? 회원가입 하러가기</a></div>
                                     </div>
                                 </div>
                             </div>
