+++ app/src/test/java/com/techcourse/DispatcherServletTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.interface21.webmvc.servlet.mvc.HandlerMapping;
-import com.techcourse.controller.LoginViewController;
+import com.techcourse.controller.legacy.LoginViewController;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
