+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/legacy/Controller.java
@@ -1,8 +1,9 @@
-package com.interface21.webmvc.servlet.mvc.asis;
+package com.interface21.webmvc.servlet.mvc.legacy;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
+@Deprecated(since = "step3", forRemoval = true)
 public interface Controller {
     String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception;
 }
