+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/legacy/ControllerHandlerAdapter.java
@@ -1,12 +1,12 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.legacy;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
+@Deprecated(since = "step3", forRemoval = true)
 public class ControllerHandlerAdapter implements HandlerAdapter {
 
     @Override
