+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/legacy/ForwardController.java
@@ -1,10 +1,11 @@
-package com.interface21.webmvc.servlet.mvc.asis;
+package com.interface21.webmvc.servlet.mvc.legacy;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.util.Objects;
 
+@Deprecated(since = "step3", forRemoval = true)
 public class ForwardController implements Controller {
 
     private final String path;
